Moved files out of src to main dir
I’m not sure why this library is not working when the files are in the
src/ dir, but they aren’t.  In the main dir, things work fine.  I
suspect that the “extern C” statements might be screwing things up.
Have to figure it out and get these back into the proper dir