Changes to debugging messages and modified the initial sync state machines
Debugging messages are now able to take a classification so that you
can see only the types of messages that you want to see.

Prepping to implement the ability to see debugging messages on remote
nodes across the mesh

Also fixed an issue in the sync status state machines that was letting
newConnectionCallback get called before the node was fully sync’d.