Notify other nodes in case of a disconnection
In message load is high, nodeSync messages are not always sent. It is assumed that if messages are coming node is connecte, and this is true.

But if other node disconnects, other nodes do not get notified and they think the disconnected node is still in the mesh.
This is an intent to always trigger nodeSyncStart in case a disconnection happens.